popup for breakwindow under GTK
this makes the window stay on top of everything else which lets it
show up on top of fullscreen windows. it has downsides. in fullscreen,
focus is weird on my system and i cant click buttons on the popup.
out of full screen that works fine but its always on top of everything
which is great for me on a tiling window manager.

anyways, trying it out, benefits outweigh the drawbacks for me.

idk how the QT backend works so im just doing it on this